Did some refactor on naming within /client to be consistent
